Fixes & Work On Function For Ordering
In this commit, I worked on some fixes / cleaning in the code I had because some pages held unnecessary code.

I also worked on the functionality of the order page. Like adding a drop down menu that allows the user to select a taco, and once ordered would give them a receipt. Currently the function is still a work in progress